moove stories of Input Forms
